Delta-X
HSR Motors is a car dealership which advertises their business on multiple platforms such as Facebook, Twitter, Google, etc. Their primary goal is to get leads from people who are interested in buying a car. 
To manage the leads that they receive, they are currently using a spreadsheet software. However, the spreadsheet software does not help them due to the large volume and the need for a lot of manual data entry.
Their lead management team typically consists of these types of users:
Call Center / Tele-calling team
Goal: To reach out to the leads on phone and understand the customer’s needs.
Sales Assistant
Goal: Be able to talk to qualified customers who are really interested in purchasing a car. 
Business Manager
Goal: Be able to get a quick-view dashboard to understand and analyze the leads data received.
After-Sales Specialist
Goal: Have a clear view of the customer’s journey so that they can contact the customer for after-sales offers like “Free general service this weekend”.
Design a web application for desktop screens that will be a one-stop solution for the lead management team at HSR Motors.